+++ src/templates.ts
@@ -64,6 +64,8 @@ export function getTemplates(options: ModuleOptions, uiConfig: Record<string, an
           if (isDev) {
             const templatePath = fileURLToPath(new URL(`./theme/${path ? `${path}/` : ''}${kebabCase(component)}`, import.meta.url))
             const themeUtilsPath = fileURLToPath(new URL('./utils/theme', import.meta.url))
+            // const defaultVariantsJson = 'undefined'
+            const prefixJson = 'undefined'
 
             return [
               `import template from ${JSON.stringify(templatePath)}`,
@@ -72,8 +74,8 @@ export function getTemplates(options: ModuleOptions, uiConfig: Record<string, an
               ...generateVariantDeclarations(variants),
               `const options = ${JSON.stringify(options, null, 2)}`,
               `let result = typeof template === 'function' ? (template as Function)(options) : template`,
-              // `result = applyDefaultVariants(result, options.theme?.defaultVariants)`,
-              `result = applyPrefixToObject(result, options.theme?.prefix)`,
+              // `result = applyDefaultVariants(result, ${defaultVariantsJson})`,
+              `result = applyPrefixToObject(result, ${prefixJson})`,
               `const theme = ${json}`,
               `export default result as typeof theme`
             ].join('\n\n')